third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-4/disable-last-property-without-semicolon-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Verifies that formatter adds a semicolon when enabling property.


Running: initFormattedStyle

Running: testFormattedDisableLast
raw cssText:
{
          color: red;
          /* margin: 0 */
      }

Running: testFormattedInsertEnd
raw cssText:
{
          color: red;
          /* margin: 0 */
          endProperty: endValue;
      }

Running: testFormattedEnable
raw cssText:
{
          color: red;
          margin: 0;
          endProperty: endValue;
      }
